Basic map rendering and scrolling
Instead of seamlessly scrolling across multiple sectors, we'll
scroll across a single sector and "cut" when the player crosses
the boundary. This is far simpler, although still ugly.

I've found that ISR_NAKED doesn't play well with optimizations,
inlining in particular. Unfortunately, the alternatives waste
too many (a dozen or so) cycles at a time when there aren't any to
spare.