Implement inertial player movement, collisions, and camera tracking
A component position is contained in three bytes: high, low, and
subpixel. High and low are the /12 %12 pair of the true position,
and subpixel is essentially a fixed point value between 0-1.

Collisions are pretty simple. Since the maximum speed in the game
is 1px/frame, every time the player's position changes, we check
whether the new position (as determined by the sprite's four
corners) is within a colliding tile (whether or not a tile is a
collider is determined by its index). If so, we do not move to the
new position.

Camera movement is simple too. As long as the player is within the
view plus some margin, the camera is stationary. If the player
approaches the view edges, the camera moves to keep him within the
margin UNLESS the view edges coincide with the sector edges.

Also deleted some old testing code.