Use an 8 bit unsigned integer for a coordinate component
Previously, all coordinates were specified by two 8 bit unsigned
values: a high value (x / 12) and a low value (x % 12). This was
a remnant from a previous plan with a much larger scrolling map.

Also fixed a bug around camera movement, and generalized several
player subroutines to work for NCPs as well.

This change simplifies a lot of calculations and reduces code size by
~600 bytes and memory usage by 20 bytes! It also could lead to
slightly reduced stack size. The only disadvantages is that the
largest possible sector size is 21x21. But that should be sufficient,
and is slightly larger than the current size.