Reduce peak stack size by 12 bytes
In main.asm, replaced calls with jumps. This is reasonable, since
the mode-specific update subroutines can return to the same place
every time. In render.asm, replaced sub-subroutines with a small
amount of duplicate code. In explore.asm, stopped unnecessarily
preserving r14-r17 and Y. Of course this violates the ABI, but it's
no big deal because render_game is a top-level subroutine.