Rework collision response
The main idea is, the goal really isn't to accurately simulate
rigid body collisions. Instead, we mainly want to ensure that enemies
and the player don't overlap and allow enemies to obstruct the
player. The first part can be accomplished with the previous force-
field effects, and the second by increasing friction when colliding
with enemies.