Implement enemy drops and player XP gains
One problem that I hadn't though about is the fact that items are
always rendered before NPCs, including the corpse NPC. As a result,
the corpses can completely obscure the dropped items. Discussed this
a bit, eventually adjusted dropped position to have a few pixels
visible as a hint, and I'm calling this a feature, it's a "looting"
mechanic. So there it is. I'm also not sure how balanced the XP gain is.