Rework animation calculations and names to replace attack2 with dash
Steven and I talked about this a bit and he suggested using a dash
instead of the secondary attack. I'd considered a dash for the
halfling, but applying it everywhere was a good idea, and it means
less class-specific code! Of course, I'd been planning to use the
secondary attack to differentiate classes a bit, particularly to
improve the mage, but mage-specific weapons will do just as well.

I also increased attack frame count to four. I'm not sure this'll
be really beneficial, but it's easy to revert and is still four net
frames fewer than before.