Indicate attack and dash cooldown time remaining
Also started using a much more graduated healing interval. Previously
it was restricted to powers of two, for easy modulus, but the new
division routines allow extending to arbitrary numbers. Also reduced
the player effect limit to 2. The idea is that even if one slot is
taken up by Curse of Ulimar or some other long-lived effect, the
other is still usable.