Improve collision system enormously
Previously, a tile was either a collider or not. This worked but
walking around could be annoying, since the corners of these tiles
would force you to change directions a lot. Also, parts of the map
which appeared to be accessible could be actually blocked by 8-
connected tiles. This change adds four 45-degree tiles.