Use a minimal AVR model during emulation
On my fairly weak laptop, this decreases "scripting"/"total" time
on Chrome from 99.9% to 98%, and dropped frames from 1 every 8-10
to 0. There's still a lot of performance left on the table, I can
easily get it to 80% by removing support for PWM, more than one
comparator, and SPM, but for now this is sufficient. I'll do that
for the release though.