Separate characters from classes
Previously, a player had a class and character, but they were
supposed to be always the same. Thus a woman was always a mage, a
man always a paladin, a halfling always a halfling. Now the character
is essentially just a sprite, and the stats depend only on class
choice. Course, not as hardcore, but who are we kidding?