Tweak upgrades and fix register clobbering
Stumbled over the bug during a test mage run when I shot a fireball
with the wooden staff, then drank a potion and switched to a melee
weapon before the fireball hit any NPCs. This exposed a bug where
I was clobbering the counter register during npc_resolve_ranged_damage.