Fix race condition
EEPROM writes must be started at most 4 cycles after setting the
EEMPE bit. But if the audio interrupt occurred at the wrong time,
the write would fail, and saving would be stuck in an incomplete
state. Strangely, this only seemed to be a problem after starting
a conversation, not really sure why.