Skip VGA generation for PC build
Can just barely achieve a solid 60 FPS on WASM with VGA generation,
but performance doubles without it. Of course, this means that the
PC build is much less accurate, but since I'm already faking audio
and the controller, I'll concede a bit more. If you want accurate
simulation, use simulate_full.c, which accurately emulates audio,
the controller, and a VGA signal.